Handle apply redux toolkit
