+++ src/App.tsx
@@ -1,36 +1,36 @@
-import React, { Fragment } from "react";
+import React, {Fragment} from "react";
 import Switch from "react-bootstrap/esm/Switch";
-import { BrowserRouter as Router, Route } from "react-router-dom";
+import {BrowserRouter as Router, Route} from "react-router-dom";
 import "./App.scss";
-import { AddTodoList } from "./components/addTodoList/addTodoList";
-import { EditTodoList } from "./components/editTodoList/editTodoList";
-import { TodoList } from "./components/todoList/todoList";
+import {AddTodoList} from "./components/addTodoList/AddTodoList";
+import {EditTodoList} from "./components/editTodoList/EditTodoList";
+import {TodoList} from "./components/todoList/TodoList";
 
 function App() {
-  return (
-    <div>
-      <Router>
-        <div className="layout">
-          <h2>Todos</h2>
-          <div className="content-todo">
-            <Switch>
-              <Route
-                exact
-                path="/"
-                render={() => (
-                  <Fragment>
-                    <AddTodoList />
-                    <TodoList />
-                  </Fragment>
-                )}
-              />
-              <Route path="/edit/:id" component={EditTodoList} />
-            </Switch>
-          </div>
+    return (
+        <div>
+            <Router>
+                <div className="layout">
+                    <h2>Todos</h2>
+                    <div className="content-todo">
+                        <Switch>
+                            <Route
+                                exact
+                                path="/"
+                                render={() => (
+                                    <Fragment>
+                                        <AddTodoList/>
+                                        <TodoList/>
+                                    </Fragment>
+                                )}
+                            />
+                            <Route path="/edit/:id" component={EditTodoList}/>
+                        </Switch>
+                    </div>
+                </div>
+            </Router>
         </div>
-      </Router>
-    </div>
-  );
+    );
 }
 
 export default App;
