+++ src/reducers/index.ts
@@ -1,6 +1,7 @@
-import { combineReducers } from 'redux'
- import todoReducer from './todoReducer';
-const rootReducer =  combineReducers({
-    todoReducer
+import {combineReducers} from 'redux'
+import todoReducer from '../components/todoList/todoSlice'
+
+const rootReducer = combineReducers({
+    todos: todoReducer
 });
 export default rootReducer;
