+++ src/style/common/style.scss
@@ -1,3 +1,3 @@
 @import "./mixis.scss";
 @import "./common.scss";
-@import "./variable.scss"
+@import "./variable.scss";
